+++ apps/e2e/tests/backend/endpoints/api/v1/team-memberships.test.ts
@@ -68,19 +68,9 @@ it("creates a team and manage users on the server", async ({ expect }) => {
         "is_paginated": false,
         "items": [
           {
-            "auth_methods": [
-              {
-                "contact_channel": {
-                  "email": "<stripped UUID>@stack-generated.example.com",
-                  "type": "email",
-                },
-                "type": "otp",
-              },
-            ],
             "auth_with_email": true,
             "client_metadata": null,
             "client_read_only_metadata": null,
-            "connected_accounts": [],
             "display_name": null,
             "has_password": false,
             "id": "<stripped UUID>",
@@ -96,19 +86,9 @@ it("creates a team and manage users on the server", async ({ expect }) => {
             "signed_up_at_millis": <stripped field 'signed_up_at_millis'>,
           },
           {
-            "auth_methods": [
-              {
-                "contact_channel": {
-                  "email": "<stripped UUID>@stack-generated.example.com",
-                  "type": "email",
-                },
-                "type": "otp",
-              },
-            ],
             "auth_with_email": true,
             "client_metadata": null,
             "client_read_only_metadata": null,
-            "connected_accounts": [],
             "display_name": null,
             "has_password": false,
             "id": "<stripped UUID>",
@@ -154,19 +134,9 @@ it("creates a team and manage users on the server", async ({ expect }) => {
         "is_paginated": false,
         "items": [
           {
-            "auth_methods": [
-              {
-                "contact_channel": {
-                  "email": "<stripped UUID>@stack-generated.example.com",
-                  "type": "email",
-                },
-                "type": "otp",
-              },
-            ],
             "auth_with_email": true,
             "client_metadata": null,
             "client_read_only_metadata": null,
-            "connected_accounts": [],
             "display_name": null,
             "has_password": false,
             "id": "<stripped UUID>",
@@ -302,6 +272,125 @@ it("removes user from team on the client", async ({ expect }) => {
     NiceResponse {
       "status": 200,
       "body": { "success": true },
+      "headers": Headers { <some fields may have been hidden> },
+    }
+  `);
+});
+
+it("creates a team and not add the current user as a member on the client", async ({ expect }) => {
+  const { userId } = await Auth.Otp.signIn();
+  const response = await niceBackendFetch("/api/v1/teams", {
+    accessType: "client",
+    method: "POST",
+    body: {
+      display_name: "My Team",
+    },
+  });
+  expect(response).toMatchInlineSnapshot(`
+    NiceResponse {
+      "status": 201,
+      "body": {
+        "client_metadata": null,
+        "client_read_only_metadata": null,
+        "display_name": "My Team",
+        "id": "<stripped UUID>",
+        "profile_image_url": null,
+      },
+      "headers": Headers { <some fields may have been hidden> },
+    }
+  `);
+
+  const response2 = await niceBackendFetch(`/api/v1/teams?user_id=me`, {
+    accessType: "client",
+    method: "GET",
+  });
+  expect(response2).toMatchInlineSnapshot(`
+    NiceResponse {
+      "status": 200,
+      "body": {
+        "is_paginated": false,
+        "items": [],
+      },
+      "headers": Headers { <some fields may have been hidden> },
+    }
+  `);
+});
+
+it("creates a team on the server and add a different user as the creator", async ({ expect }) => {
+  const user1Mailbox = createMailbox();
+  backendContext.set({ mailbox: user1Mailbox });
+  const { userId: userId1 } = await Auth.Otp.signIn();
+  backendContext.set({ mailbox: createMailbox() });
+  await Auth.Otp.signIn();
+
+  const response = await niceBackendFetch("/api/v1/teams", {
+    accessType: "server",
+    method: "POST",
+    body: {
+      display_name: "My Team",
+      creator_user_id: userId1,
+    },
+  });
+  expect(response).toMatchInlineSnapshot(`
+    NiceResponse {
+      "status": 201,
+      "body": {
+        "client_metadata": null,
+        "client_read_only_metadata": null,
+        "created_at_millis": <stripped field 'created_at_millis'>,
+        "display_name": "My Team",
+        "id": "<stripped UUID>",
+        "profile_image_url": null,
+        "server_metadata": null,
+      },
+      "headers": Headers { <some fields may have been hidden> },
+    }
+  `);
+
+  backendContext.set({ mailbox: user1Mailbox });
+  await Auth.Otp.signIn();
+
+  const response2 = await niceBackendFetch(`/api/v1/teams?user_id=me`, {
+    accessType: "client",
+    method: "GET",
+  });
+  expect(response2).toMatchInlineSnapshot(`
+    NiceResponse {
+      "status": 200,
+      "body": {
+        "is_paginated": false,
+        "items": [
+          {
+            "client_metadata": null,
+            "client_read_only_metadata": null,
+            "display_name": "My Team",
+            "id": "<stripped UUID>",
+            "profile_image_url": null,
+          },
+        ],
+      },
+      "headers": Headers { <some fields may have been hidden> },
+    }
+  `);
+});
+
+it("is not allowed to create a team and add a different user as the creator on the client", async ({ expect }) => {
+  const { userId: userId1 } = await Auth.Otp.signIn();
+  backendContext.set({ mailbox: createMailbox() });
+  await Auth.Otp.signIn();
+
+  const response = await niceBackendFetch("/api/v1/teams", {
+    accessType: "client",
+    method: "POST",
+    body: {
+      display_name: "My Team",
+      creator_user_id: userId1,
+    },
+  });
+  expect(response).toMatchInlineSnapshot(`
+    NiceResponse {
+      "status": 403,
+      "body": "You cannot add a user to the team as the creator that is not yourself on the client.",
       "headers": Headers { <some fields may have been hidden> },
     }
   `);