+++ apps/e2e/tests/backend/endpoints/api/v1/users.test.ts
@@ -77,19 +77,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -117,19 +107,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -202,19 +182,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -241,19 +211,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": { "key": "value" },
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -390,19 +350,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -429,19 +379,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -566,19 +506,9 @@ describe("with client access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": { "key": "value" },
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -682,19 +612,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -727,19 +647,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -774,7 +684,43 @@ describe("with server access", () => {
     `);
   });
 
-  it.todo("should be able to list users");
+  it("should be able to list users", async ({ expect }) => {
+    await Project.createAndSwitch({ config: { magic_link_enabled: true } });
+    await Auth.Otp.signIn();
+
+    const response = await niceBackendFetch("/api/v1/users", {
+      accessType: "server",
+    });
+    expect(response).toMatchInlineSnapshot(`
+      NiceResponse {
+        "status": 200,
+        "body": {
+          "is_paginated": false,
+          "items": [
+            {
+              "auth_with_email": true,
+              "client_metadata": null,
+              "client_read_only_metadata": null,
+              "display_name": null,
+              "has_password": false,
+              "id": "<stripped UUID>",
+              "last_active_at_millis": <stripped field 'last_active_at_millis'>,
+              "oauth_providers": [],
+              "primary_email": "<stripped UUID>@stack-generated.example.com",
+              "primary_email_verified": true,
+              "profile_image_url": null,
+              "requires_totp_mfa": false,
+              "selected_team": null,
+              "selected_team_id": null,
+              "server_metadata": null,
+              "signed_up_at_millis": <stripped field 'signed_up_at_millis'>,
+            },
+          ],
+        },
+        "headers": Headers { <some fields may have been hidden> },
+      }
+    `);
+  });
 
   it("should be able to read a user", async ({ expect }) => {
     await Auth.Otp.signIn();
@@ -792,19 +738,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -835,11 +771,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [],
           "auth_with_email": false,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -874,19 +808,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Dough",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -921,23 +845,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-            {
-              "identifier": "<stripped UUID>@stack-generated.example.com",
-              "type": "password",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": true,
           "id": "<stripped UUID>",
@@ -1017,19 +927,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1082,11 +982,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [],
           "auth_with_email": false,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1118,23 +1016,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-            {
-              "identifier": "<stripped UUID>@stack-generated.example.com",
-              "type": "password",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": true,
           "id": "<stripped UUID>",
@@ -1181,23 +1065,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-            {
-              "identifier": "<stripped UUID>@stack-generated.example.com",
-              "type": "password",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": true,
           "id": "<stripped UUID>",
@@ -1226,11 +1096,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 201,
         "body": {
-          "auth_methods": [],
           "auth_with_email": false,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1283,19 +1151,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1320,19 +1178,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1366,19 +1214,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": "John Doe",
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1412,23 +1250,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-            {
-              "identifier": "<stripped UUID>@stack-generated.example.com",
-              "type": "password",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": null,
           "client_read_only_metadata": null,
-          "connected_accounts": [],
           "display_name": null,
           "has_password": true,
           "id": "<stripped UUID>",
@@ -1511,19 +1335,9 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "auth_methods": [
-            {
-              "contact_channel": {
-                "email": "<stripped UUID>@stack-generated.example.com",
-                "type": "email",
-              },
-              "type": "otp",
-            },
-          ],
           "auth_with_email": true,
           "client_metadata": { "key": "client value" },
           "client_read_only_metadata": { "key": "client read only value" },
-          "connected_accounts": [],
           "display_name": null,
           "has_password": false,
           "id": "<stripped UUID>",
@@ -1553,5 +1367,40 @@ describe("with server access", () => {
       },
     });
     expect(response.body.profile_image_url).toEqual("http://localhost:8101/open-graph-image.png");
+  });
+
+  it("should be able to update primary email", async ({ expect }) => {
+    await Auth.Otp.signIn();
+    const response = await niceBackendFetch("/api/v1/users/me", {
+      accessType: "server",
+      method: "PATCH",
+      body: {
+        primary_email: "new-primary-email@example.com",
+      },
+    });
+    expect(response).toMatchInlineSnapshot(`
+      NiceResponse {
+        "status": 200,
+        "body": {
+          "auth_with_email": true,
+          "client_metadata": null,
+          "client_read_only_metadata": null,
+          "display_name": null,
+          "has_password": false,
+          "id": "<stripped UUID>",
+          "last_active_at_millis": <stripped field 'last_active_at_millis'>,
+          "oauth_providers": [],
+          "primary_email": "new-primary-email@example.com",
+          "primary_email_verified": true,
+          "profile_image_url": null,
+          "requires_totp_mfa": false,
+          "selected_team": null,
+          "selected_team_id": null,
+          "server_metadata": null,
+          "signed_up_at_millis": <stripped field 'signed_up_at_millis'>,
+        },
+        "headers": Headers { <some fields may have been hidden> },
+      }
+    `);
   });
 });